move __CENSORED__ key to privacyidea.lib.crypto
