+++ privacyidea/lib/machineresolver.py
@@ -36,9 +36,8 @@ from sqlalchemy import func, select
 from privacyidea.lib.config import get_machine_resolver_class_dict
 from privacyidea.lib.utils import (sanity_name_check, get_data_from_params, fetch_one_resource)
 from privacyidea.lib.utils.export import (register_import, register_export)
-from .crypto import encryptPassword, decryptPassword
+from .crypto import encryptPassword, decryptPassword, CENSORED
 from .log import log_with
-from .resolver import CENSORED
 from privacyidea.lib.params import get_required
 from ..models import (MachineResolver,
                       MachineResolverConfig, db)
