+++ privacyidea/lib/resolver.py
@@ -55,14 +55,13 @@ from privacyidea.lib.utils import (sanity_name_check, get_data_from_params,
                                    is_true)
 from privacyidea.lib.utils.export import (register_import, register_export)
 from .config import (get_resolver_types, get_resolver_classes, get_config_object)
-from .crypto import encryptPassword
+from .crypto import encryptPassword, CENSORED
 from .error import ConfigAdminError
 from .log import log_with
 from privacyidea.lib.params import get_required
 from ..models import (Resolver,
                       ResolverConfig, save_config_timestamp, db)
 
-CENSORED = "__CENSORED__"
 log = logging.getLogger(__name__)
