+++ privacyidea/lib/smtpserver.py
@@ -29,8 +29,7 @@ from flask import current_app
 from sqlalchemy import select
 
 from privacyidea.lib.crypto import (decryptPassword, encryptPassword,
-                                    FAILED_TO_DECRYPT_PASSWORD)
-from privacyidea.lib.log import log_with
+                                    FAILED_TO_DECRYPT_PASSWORD, CENSORED)
 from privacyidea.lib.resolver import CENSORED
 from privacyidea.lib.queue import job, wrap_job, has_job_queue
 from privacyidea.lib.utils import fetch_one_resource, to_unicode
